--- micro_services/ApiGateWay/requirements.txt
@@ -1,3 +1,0 @@
-grpcio==1.54.0
-grpcio-tools==1.49.0
-protobuf==4.22.3
